Remove TreeViewer constructor with projectKey parameter
