+++ src/test/java/de/uhd/ifi/se/decision/management/jira/rest/viewrest/TestGetTreeViewer.java
@@ -2,50 +2,89 @@ package de.uhd.ifi.se.decision.management.jira.rest.viewrest;
 
 import static org.junit.Assert.assertEquals;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.ws.rs.core.Response.Status;
 
 import org.junit.Before;
 import org.junit.Test;
 
+import com.atlassian.jira.mock.servlet.MockHttpServletRequest;
+import com.atlassian.jira.user.ApplicationUser;
+
 import de.uhd.ifi.se.decision.management.jira.TestSetUp;
+import de.uhd.ifi.se.decision.management.jira.filtering.FilterSettings;
 import de.uhd.ifi.se.decision.management.jira.rest.ViewRest;
-import net.java.ao.test.jdbc.NonTransactional;
+import de.uhd.ifi.se.decision.management.jira.testdata.JiraUsers;
 
 public class TestGetTreeViewer extends TestSetUp {
 	private ViewRest viewRest;
+	private HttpServletRequest request;
+	private FilterSettings filterSettings;
 
 	@Before
 	public void setUp() {
-		viewRest = new ViewRest();
 		init();
+		viewRest = new ViewRest();
+		request = new MockHttpServletRequest();
+		ApplicationUser user = JiraUsers.SYS_ADMIN.getApplicationUser();
+		request.setAttribute("user", user);
+		filterSettings = new FilterSettings("TEST", "");
 	}
 
-	@Test
-	@NonTransactional
-	public void testProjectKeyNullKnowledgeTypeNull() {
-		assertEquals(Status.BAD_REQUEST.getStatusCode(), viewRest.getTreeViewer(null, null).getStatus());
-	}
+	// @Test
+	// @NonTransactional
+	// public void testProjectKeyNullKnowledgeTypeNull() {
+	// assertEquals(Status.BAD_REQUEST.getStatusCode(), viewRest.getTreeViewer(null,
+	// null).getStatus());
+	// }
+	//
+	// @Test
+	// @NonTransactional
+	// public void testProjectKeyNonExistentKnowledgeTypeNull() {
+	// assertEquals(Status.BAD_REQUEST.getStatusCode(),
+	// viewRest.getTreeViewer("NotExistingProjectKey", null).getStatus());
+	// }
+	//
+	// @Test
+	// @NonTransactional
+	// public void testProjectKeyExistentKnowledgeTypeNull() {
+	// assertEquals(200, viewRest.getTreeViewer("TEST", null).getStatus());
+	// }
+	//
+	// @Test
+	// public void testProjectKeyExistentKnowledgeTypeEmpty() {
+	// assertEquals(200, viewRest.getTreeViewer("TEST", "").getStatus());
+	// }
+	//
+	// @Test
+	// public void testProjectKeyExistentKnowledgeTypeFilled() {
+	// assertEquals(200, viewRest.getTreeViewer("TEST", "Issue").getStatus());
+	// }
 
 	@Test
-	@NonTransactional
-	public void testProjectKeyNonExistentKnowledgeTypeNull() {
+	public void testRequestNullJiraIssueKeyNullFilterSettingsNull() {
 		assertEquals(Status.BAD_REQUEST.getStatusCode(),
-				viewRest.getTreeViewer("NotExistingProjectKey", null).getStatus());
+				viewRest.getTreeViewer(null, null).getStatus());
 	}
 
 	@Test
-	@NonTransactional
-	public void testProjectKeyExistentKnowledgeTypeNull() {
-		assertEquals(200, viewRest.getTreeViewer("TEST", null).getStatus());
+	public void testRequestValidFilterSettingsValid() {
+		assertEquals(Status.OK.getStatusCode(),
+				viewRest.getTreeViewer(request, filterSettings).getStatus());
 	}
 
 	@Test
-	public void testProjectKeyExistentKnowledgeTypeEmpty() {
-		assertEquals(200, viewRest.getTreeViewer("TEST", "").getStatus());
+	public void testRequestValidJiraIssueKeyValidFilterSettingsValid() {
+		filterSettings.setSelectedElement("TEST-1");
+		assertEquals(Status.OK.getStatusCode(),
+				viewRest.getTreeViewer(request, filterSettings).getStatus());
 	}
 
 	@Test
-	public void testProjectKeyExistentKnowledgeTypeFilled() {
-		assertEquals(200, viewRest.getTreeViewer("TEST", "Issue").getStatus());
+	public void testRequestValidJiraIssueKeyValidFilterSettingsInvalid() {
+		FilterSettings filterSettings = new FilterSettings("NOTEXISTINGPROJECTKEY", "");
+		filterSettings.setSelectedElement("NOTEXISTINGPROJECTKEY-5");
+		assertEquals(Status.BAD_REQUEST.getStatusCode(),
+				viewRest.getTreeViewer(request, filterSettings).getStatus());
 	}
 }
